Split unit test
